refactor: change structure of the project and paths naming
