+++ src/routes/AppRouter.tsx
@@ -1,8 +1,8 @@
-import { AppRoutes } from '../enum/AppRoutes';
-import AboutPage from '../pages/AboutPage';
-import HomePage from '../pages/HomePage';
-import NotFoundPage from '../pages/NotFoundPage';
-import { Routes, Route } from 'react-router-dom';
+import { Route, Routes } from 'react-router-dom';
+import AboutPage from 'src/pages/AboutPage';
+import { AppRoutes } from 'src/enum/AppRoutes';
+import HomePage from 'src/pages/HomePage';
+import NotFoundPage from 'src/pages/NotFoundPage';
 
 const AppRouter = () => {
   return (
