+++ tsconfig.json
@@ -12,7 +12,10 @@
     "forceConsistentCasingInFileNames": true,
     "strict": true,
     "skipLibCheck": true,
-    "baseUrl": "./src",
+    "baseUrl": ".",
+    "paths": {
+      "src/*": ["src/*"]
+    },
     "resolveJsonModule": true
   },
   "include": ["vite.config.ts", "./src/**/*", "styled.d.ts"]
