+++ vite.config.ts
@@ -7,6 +7,11 @@ import { defineConfig } from 'vite';
 // https://vitejs.dev/config/
 export default defineConfig({
   plugins: [react()],
+  resolve: {
+    alias: {
+      src: '/src',
+    },
+  },
   test: {
     globals: true,
     environment: 'jsdom',
@@ -25,6 +30,7 @@ export default defineConfig({
         'src/enum',
         'src/routes',
         'src/types',
+        '**/index.ts',
       ],
     },
   },
